fh add: add dry run support

+++ src/cli/cmd/add/mod.rs
@@ -37,6 +37,9 @@ pub(crate) struct AddSubcommand {
     /// Whether to insert a new input at the top of or the bottom of an existing `inputs` attrset.
     #[clap(long, default_value_t = InputsInsertionLocation::Top)]
     pub(crate) insertion_location: InputsInsertionLocation,
+    /// Print to stdout the new flake.nix contents instead of writing it to disk.
+    #[clap(long)]
+    pub(crate) dry_run: bool,
 
     #[clap(from_global)]
     api_addr: url::Url,
@@ -65,7 +68,11 @@ impl CommandExecute for AddSubcommand {
             self.insertion_location,
         )?;
 
-        tokio::fs::write(self.flake_path, new_flake_contents).await?;
+        if self.dry_run {
+            println!("{new_flake_contents}");
+        } else {
+            tokio::fs::write(self.flake_path, new_flake_contents).await?;
+        }
 
         Ok(ExitCode::SUCCESS)
     }
